README.md
If you're looking to start trading cryptocurrencies with a leading platform while maximizing bonuses and discounts, the BingX referral code IMOIJ1 is your gateway to exciting rewards in 2025. Whether you're a new user or planning to sign up soon, this BingX signup bonus can help you claim up to 5125 USDT in rewards alongside a 50% discount on trading fees. Read on to discover how to use this BingX discount code, the latest updates, and all the details you need to get started.